Organizaed the usecases or organization and added securityclearances

+++ models/lib/use_cases/organization_use_cases/remove_member_from_organization.dart
@@ -1,4 +1,5 @@
 import 'package:models/models/member.dart';
+import 'package:models/models/role.dart';
 
 ///
 /// remove_member_from_organization.dart
@@ -8,19 +9,30 @@ import 'package:models/models/member.dart';
 /// Copyright 2021 Po. All rights reserved.
 ///
 
-mixin RemoveMemberFromOrganization {
+mixin RemoveMemberFromOrganizationUseCase {
+  /// Removes a member from the organization
+  ///
+  /// [memberID] id of the member being removed from the organization
+  /// [accessorID] id of the member that is attempting to remove [memberId]
+  /// Returns a [ResultValue] if successful else returns [ResultError]
   @required
-  Future<Result> execute(MemberID memberID);
+  Future<Result> execute(
+      {required MemberID accessorId, required MemberID memberID});
 }
 
-class DefaultRemoveMemberFromOrganization
-    implements RemoveMemberFromOrganization {
+class DefaultRemoveMemberFromOrganizationUseCase
+    implements RemoveMemberFromOrganizationUseCase {
   MemberRepository _memberRepository;
 
-  DefaultRemoveMemberFromOrganization(this._memberRepository);
+  DefaultRemoveMemberFromOrganizationUseCase(this._memberRepository);
 
   @override
-  Future<Result> execute(MemberID memberID) async {
+  Future<Result> execute(
+      {required MemberID accessorId, required MemberID memberID}) async {
+    Member accessor = await _memberRepository.find(accessorId);
+    if (accessor.role.securityClearance < SecurityClearance.level1)
+      return Result.error("Access to Remove Member Denied");
+
     Result result = await _memberRepository.remove(memberID);
     return result;
   }
